requirements.txt.txt
Pillow = 3.2.0
selenium = 2.53.6